benchmark/requirements.txt
google-cloud==0.34.0
fabric==3.0.0
matplotlib==3.3.4
google-cloud-compute==1.18.0